refactor: 인물 도메인 모델 수정 (#27)

+++ feature/screens/src/main/java/com/example/metasearch/feature/screens/Screens.kt
@@ -20,7 +20,7 @@ data object NLSearchScreen : Screen
 
 @Parcelize
 data class PersonDetailScreen(
-    val personId: Int,
+    val personId: Long,
 ) : Screen
 
 @Parcelize
